ai_detection.py
import cv2
import logging
import numpy as np
from typing import Tuple, List, Dict

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class AIDetector:
    def __init__(self):
        """Initialize the AI detector."""
        self.confidence_threshold = 0.5

    def process_frame(
        self, frame: np.ndarray
    ) -> Tuple[List[Dict], np.ndarray]:
        """Process a frame and return detections and annotated frame."""
        try:
            # For now, just return empty detections and the original frame
            return [], frame

        except Exception as e:
            logger.error(f"Error processing frame: {str(e)}")
            return [], frame

    def process_image_file(
        self, image_path: str
    ) -> Tuple[List[Dict], np.ndarray]:
        """Process an image file and return detections and annotated image."""
        try:
            # Read image
            frame = cv2.imread(image_path)
            if frame is None:
                raise Exception(f"Failed to read image: {image_path}")
            
            # Process frame
            return self.process_frame(frame)
            
        except Exception as e:
            logger.error(f"Error processing image file: {str(e)}")
            return [], None

# Create a singleton instance
detector = AIDetector()