requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.1
aiofiles>=0.7.0
starlette>=0.14.2

# Database
sqlalchemy>=1.4.23
alembic>=1.7.1

# Security
passlib>=1.7.4
bcrypt>=3.2.0
python-jose>=3.3.0

# Video Processing
opencv-python>=4.5.3.56
numpy>=1.21.2

# System Monitoring
psutil>=5.8.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.19.0

# Development
black>=21.7b0
flake8>=3.9.2
isort>=5.9.3
itsdangerous
